Improvement/warning undeclared param type

Approved-by: Jose Rodriguez <boriel@gmail.com>

+++ tests/functional/param3.bas
@@ -1,6 +1,6 @@
 ' Declare a parameter of type string using $ sigil suffix
 
-declare function test(s$)
+declare function test(s$) as float
 
 function test$(s$)
     return s$ + "A"
